% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-graphs.R
\name{causal_graph_tikz}
\alias{causal_graph_tikz}
\title{Tikz for Causal Graph}
\usage{
causal_graph_tikz(adj_mat, colour_nodes = "yellow!60")
}
\arguments{
\item{adj_mat}{Adjacency matrix, with named columns and rows.}

\item{colour_nodes}{Colour of nodes. Default to "yellow!60". If specific
colour is desired for a node, a vector with colours in the same order
as the column names must be provided.}
}
\description{
Tikz for Causal Graph
}
