% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-graphs.R
\name{generate_coordinates_graph}
\alias{generate_coordinates_graph}
\title{Generates the coordinates used to plot the causal graph from an adjacency
matrix}
\usage{
generate_coordinates_graph(adj_mat)
}
\arguments{
\item{adj_mat}{Adjacency matrix, with named columns and rows.}
}
\value{
A list with three elements:
* `coordinates`: Cordinates of the nodes.
* `edges`: Edges.
* `name_nodes`: Names of the nodes.
}
\description{
Generates the coordinates used to plot the causal graph from an adjacency
matrix
}
