% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-graphs.R
\name{generate_edges}
\alias{generate_edges}
\title{Function to generate TikZ code for edges}
\usage{
generate_edges(edges)
}
\arguments{
\item{edges}{Index of edges for each node.}
}
\description{
Function to generate TikZ code for edges
}
