% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_bin_classif}
\alias{metrics_bin_classif}
\title{Metrics for a Scoring Classifier}
\usage{
metrics_bin_classif(obs, pred, threshold = 0.5)
}
\arguments{
\item{obs}{Vector of observed binary (0/1) variable.}

\item{pred}{Vector of predicted scores.}

\item{threshold}{Threshold above which predicted scores are classified as 1.}
}
\value{
A named numeric vector with the following metrics:
\itemize{
\item \code{mean_pred}: Mean predicted scores
\item \code{n_obs}: Number of observations
\item \code{TP}: Count of true positives
\item \code{FP}: Count of false positives
\item \code{FN}: Count of false negatives
\item \code{TN}: Count of true negatives
\item \code{P}: Count of positives (\code{TP} + \code{FN})
\item \code{N}: Count of positives (\code{FP} + \code{TN})
\item \code{TPR}: True positive rate (\code{TP} / \code{P})
\item \code{FPR}: False positive rate (\code{FP} / \code{N})
\item \code{TNR}: True negative rate (\code{TN} / \code{N})
\item \code{FNR}: False negative rate (\code{FN} / \code{P})
}
}
\description{
From predicted scores of a binary scoring classifier and corresponding
observed outcome (\eqn{Y\in\{0,1\}}), computes various metrics. The scoring
classifier is turned into a binary classifier by setting a threshold above
which individuals are predicted as 1 (or as the second level of the binary
outcome if it is a factor variable).
}
\examples{
obs <-  c( 0,  0,  0,  0,  1,  1,  1)
pred <- c(.2, .3, .8, .1, .6, .5, .1)
metrics_bin_classif(obs, pred, threshold =  .5)
}
