% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.sequential_transport}
\alias{predict.sequential_transport}
\title{Transported Values by Sequential Transport}
\usage{
\method{predict}{sequential_transport}(object, newdata, data, ...)
}
\arguments{
\item{object}{Object of class \code{"sequential_transport"}.}

\item{newdata}{An optional data frame/tibble in which to look for variables
with which to predict.}

\item{data}{Initial data on which the object has been created.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Transported Values by Sequential Transport
}
\examples{
sim_dat <- simul_dataset()
# Causal graph:
variables <- c("S", "X1", "X2", "Y")
adj <- matrix(
  # S  X1 X2 Y
  c(0, 1, 1, 1,# S
    0, 0, 1, 1,# X1
    0, 0, 0, 1,# X2
    0, 0, 0, 0  # Y
  ),
  ncol = length(variables),
  dimnames = rep(list(variables), 2),
  byrow = TRUE
)
# To visualize the causal graph:
# causal_graph <- fairadapt::graphModel(adj)
# plot(causal_graph)

# Sequential transport according to the causal graph
transported <- seq_trans(data = sim_dat, adj = adj, s = "S", S_0 = 0, y = "Y")
# The transported values for group in S=0:
predict(transported)
# Transport of new observations from S=0:
new_obs <- data.frame(S=0, X1 = c(-1, -1.5), X2 = c(0, -.5))
predict(transported, newdata = new_obs, data = sim_dat)
}
