% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-transport.R
\name{seq_trans}
\alias{seq_trans}
\title{Sequential Transport Using a Pre-Defined Causal Graph}
\usage{
seq_trans(data, adj, s, S_0, y, num_neighbors = 5, silent = FALSE)
}
\arguments{
\item{data}{Data frame with the observations.}

\item{adj}{Adjacency matrix for the causal graph.}

\item{s}{Name of the sensitive attribute column in the data.}

\item{S_0}{Label of the sensitive attribute in the source distribution.}

\item{y}{Name of the outcome variable in the data.}

\item{num_neighbors}{Number of neighbors to use in the weighted quantile
estimation. Default to 5.}

\item{silent}{If \code{TRUE}, the messages showing progress in the estimation are
not shown. Default to \code{silent=FALSE}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{transported}: A named list with the transported values. The names are those of the variables.
\item \code{weights}: A list with the weights of each observation in the two groups.
\item \code{ecdf}: A list with empirical distribution functions for numerical variables.
\item \code{ecdf_values}: A list with the values of the ecdf evaluated for each observation in the source distribution.
\item \code{fit_for_categ}: A list with the estimated multinomial models to predict categories using parents characteristics
\item \code{params}: A list with some parameters used to transport observations:
\itemize{
\item \code{adj}: Adjacency matrix.
\item \code{top_order}: Topological ordering.
\item \code{s}: Name of the sensitive attribute.
\item \code{S_0}: Label of the sensitive attribute in the source distribution.
\item \code{y}: Name of the outcome variable in the data.
\item \code{num_neighbors}: Number of neighbors used when computing quantiles.
}
}
}
\description{
The sensitive attribute, S, is assumed to be a binary variable with value
$S_0$ in the source distribution and $S_1$ in the target distribution.
}
