% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-transport.R
\name{seq_trans_new}
\alias{seq_trans_new}
\title{Sequential Transport of New Observations Using a Pre-Defined Causal Graph}
\usage{
seq_trans_new(x, newdata, data)
}
\arguments{
\item{x}{Object containing previously transported values, obtained with
`seq_transport_binary()`.}

\item{newdata}{New data with observations from initial group to transport.}

\item{data}{Data used during the construction of counterfactuals with
`seq_transport_binary()`.}
}
\value{
A list with the transported values. Each element contains the
         transported values of a variable.
}
\description{
This function transports new observations from the source group to the target
group, using the mappings previously learned when evaluating the
`seq_transport_binary()` function.
}
