% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-data.R
\name{simul_dataset}
\alias{simul_dataset}
\title{Dataset Generation with a Sensitive Attribute}
\usage{
simul_dataset(
  n_0 = 100,
  n_1 = 100,
  mu_0 = rep(-1, 2),
  mu_1 = rep(1.5, 2),
  sigma_0 = matrix(c(1, 0.5, 0.5, 1) * 1.2^2, 2, 2),
  sigma_1 = matrix(c(1, -0.4, -0.4, 1) * 0.9^2, 2, 2)
)
}
\arguments{
\item{n_0}{Number of observations in group S=0 (Default to 100).}

\item{n_1}{Number of observations in group S=1 (Default to 100).}

\item{mu_0}{Mean of the bivariate Gaussian in group S=0. (Default to -1).}

\item{mu_1}{Mean of the bivariate Gaussian in group S=1. (Default to 1.5).}

\item{sigma_0}{A symmetric positive-definite matrix representing the
variance-covariance matrix of the distribuion in group S=0.}

\item{sigma_1}{A symmetric positive-definite matrix representing the
variance-covariance matrix of the distribuion in group S=1.}
}
\description{
This function generates a dataset where covariates \eqn{X = (X1, X2)} are drawn from bivariate Gaussian distributions that depend on a binary sensitive attribute \eqn{S}. The binary outcome \eqn{Y} is generated based on \eqn{X} and \eqn{S} through a logistic model.
}
\details{
For \eqn{S = 0}, \eqn{X} is sampled from a multivariate normal distribution with mean vector \eqn{[-1, -1]} and covariance matrix \deqn{\begin{bmatrix} 1.2^2 & 1.2^2 \cdot 0.5 \\ 1.2^2 \cdot 0.5 & 1.2^2 \end{bmatrix}.}

For \eqn{S = 1}, \eqn{X} is sampled from a multivariate normal distribution with mean vector \eqn{[1.5, 1.5]} and covariance matrix \deqn{\begin{bmatrix} 0.9^2 & 0.9^2 \cdot -0.4 \\ 0.9^2 \cdot -0.4 & 0.9^2 \end{bmatrix}.}

The logistic model uses a linear predictor \eqn{\eta} to determine the probability \eqn{p} of \eqn{Y = 1}:
\itemize{
\item For \eqn{S = 0}, \eqn{\eta = 0.6 \cdot X1 + 0.2 \cdot X2}.
\item For \eqn{S = 1}, \eqn{\eta = 0.4 \cdot X1 + 0.6 \cdot X2}.
}

\eqn{Y} is sampled from a Bernoulli distribution with probability \eqn{p}, where \deqn{p = \frac{\exp(\eta)}{1 + \exp(\eta)}.}

The final dataset includes \eqn{S}, \eqn{X1}, \eqn{X2}, and \eqn{Y}.
}
\examples{
sim_dat <- simul_dataset()
head(sim_dat)
}
