% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_dataset}
\alias{split_dataset}
\title{Define the function to split the dataset into train/test sets}
\usage{
split_dataset(data, seed = NULL, train_ratio = 0.7)
}
\arguments{
\item{data}{Dataset.}

\item{seed}{Seed number to use for replication, if provided.}

\item{train_ratio}{Ratio of data to use in the train set (default to 0.7).}
}
\value{
A list with two elements:
\itemize{
\item \code{data_train}: The train set.
\item \code{data_test}: The test set.
}
}
\description{
Define the function to split the dataset into train/test sets
}
\examples{
df <- simul_dataset()
sets <- split_dataset(df)
data_train <- sets$data_train
data_test <- sets$data_test
}
\seealso{
\code{\link[=simul_dataset]{simul_dataset()}}
}
