% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-transport.R
\name{topological_ordering}
\alias{topological_ordering}
\title{Topological Ordering}
\source{
This function comes from the fairadapt package. Drago Plecko,
        Nicolai Meinshausen (2020). Fair data adaptation with quantile
        preservation Journal of Machine Learning Research, 21.242, 1-44.
        URL https://www.jmlr.org/papers/v21/19-966.html.
}
\usage{
topological_ordering(adj_mat)
}
\arguments{
\item{adj_mat}{Adjacency matrix with names of the variables for both rows and
columns.}
}
\value{
A character vector (names of the variables) providing a topological
        ordering.
}
\description{
Topological Ordering
}
