% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-graphs.R
\name{causal_graph_tikz}
\alias{causal_graph_tikz}
\title{Tikz for Causal Graph}
\usage{
causal_graph_tikz(adj_mat)
}
\arguments{
\item{adj_mat}{Adjacency matrix, with named columns and rows.}
}
\description{
Tikz for Causal Graph
}
