% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{counter_fair_metrics}
\alias{counter_fair_metrics}
\title{Counterfactual Fairness Metrics for Binary Scoring Classifier}
\usage{
counter_fair_metrics(obs_0, obs_1, pred_0, pred_0_t, pred_1, threshold)
}
\arguments{
\item{obs_0}{Vector of observed binary (0/1) variable in group 0.}

\item{obs_1}{Vector of observed binary (0/1) variable in group 1.}

\item{pred_0}{Vector of predicted scores in group 0.}

\item{pred_0_t}{Vector of predicted scores computed based on counterfactual
characteristics of individuals from group 0.}

\item{pred_1}{Vector of predicted scores in group 1.}

\item{threshold}{Threshold above which predicted scores are classified as 1.}
}
\value{
A tibble with the metrics in rows in each group (in column). The
following metrics are returned:
\itemize{
\item \code{mean_pred}: Mean predicted scores
\item \code{n_obs}: Number of observations
\item \code{TP}: Count of true positives
\item \code{FP}: Count of false positives
\item \code{FN}: Count of false negatives
\item \code{TN}: Count of true negatives
\item \code{P}: Count of positives (\code{TP} + \code{FN})
\item \code{N}: Count of positives (\code{FP} + \code{TN})
\item \code{TPR}: True positive rate (\code{TP} / \code{P})
\item \code{FPR}: False positive rate (\code{FP} / \code{N})
\item \code{TNR}: True negative rate (\code{TN} / \code{N})
\item \code{FNR}: False negative rate (\code{FN} / \code{P})
}
}
\description{
From predicted scores of a binary scoring classifier in two groups (group 0,
and group 1) and corresponding observed outcome (\eqn{Y\in\{0,1\}}), as well
as predicted scores for counterfactual individuals from group 0, computes
various metrics within each group. The scoring classifier is turned into a
binary classifier by setting a threshold above which individuals are
predicted as 1 (or as the second level of the binary outcome if it is a
factor variable).
}
\examples{
obs_0 <-    c( 0,  0,  0,  0,  1,  1,  1)
pred_0 <-   c(.2, .3, .8, .1, .6, .5, .1)
pred_0_t <- c(.2, .2, .7, .6, .8, .7, .3)
obs_1 <-    c( 0,  1,  0,  1,  1,  1,  1)
pred_1 <-   c(.6, .8, .2, .8, .5, .9, .8)
counter_fair_metrics(obs_0, obs_1, pred_0, pred_0_t, pred_1, threshold =  .5)
}
