% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-graphs.R
\name{generate_nodes}
\alias{generate_nodes}
\title{Function to generate TikZ code for nodes}
\usage{
generate_nodes(coordinates, name_nodes, colour_nodes = "yellow!60")
}
\arguments{
\item{coordinates}{Coordinates of the nodes.}

\item{name_nodes}{Vector of names of the nodes.}

\item{colour_nodes}{Colour of nodes. Default to "yellow!60".}
}
\description{
Function to generate TikZ code for nodes
}
