% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-transport.R
\name{seq_trans_new}
\alias{seq_trans_new}
\title{Sequential Transport of New Observations Using a Pre-Defined Causal Graph}
\usage{
seq_trans_new(x, newdata, data)
}
\arguments{
\item{x}{Object containing previously transported values, obtained with
\code{seq_transport_binary()}.}

\item{newdata}{New data with observations from initial group to transport.}

\item{data}{Data used during the construction of counterfactuals with
\code{seq_transport_binary()}.}
}
\value{
A list with the transported values. Each element contains the
transported values of a variable.
}
\description{
This function transports new observations from the source group to the target
group, using the mappings previously learned when evaluating the
\code{seq_transport_binary()} function.
}
\examples{
# Data with two groups: S=0, S=1, an outcome Y and two covariates X1 and X2
sim_dat <- simul_dataset()
# Causal graph:
variables <- c("S", "X1", "X2", "Y")
adj <- matrix(
  # S  X1 X2 Y
  c(0, 1, 1, 1,# S
    0, 0, 1, 1,# X1
    0, 0, 0, 1,# X2
    0, 0, 0, 0  # Y
  ),
  ncol = length(variables),
  dimnames = rep(list(variables), 2),
  byrow = TRUE
)

# Sequential transport according to the causal graph
transported <- seq_trans(data = sim_dat, adj = adj, s = "S", S_0 = 0, y = "Y")
# Transported values from S=0 to S=1, using the causal graph.
transported_val <- as.data.frame(transported$transported)
head(transported_val)

# Transport of two new observations
new_obs <- data.frame(X1 = c(-2, -1), X2 = c(-1, -2), S = c(0, 0))
transport_new <- seq_trans_new(
  x = transported, newdata = new_obs, data = sim_dat
)
new_obs_transported <- as.data.frame(transport_new)
new_obs_transported
}
\seealso{
\code{\link[=seq_trans]{seq_trans()}}, \code{\link[=simul_dataset]{simul_dataset()}}
}
