% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{seqtransfairness}
\alias{seqtransfairness}
\title{seqtransfairness}
\description{
{seqtransfairness} is an R package that provides tools to create
counterfactuals based on sequential conditional transport using a causal
graph assumed to be known. This approach links two existing approaches
to derive counterfactuals: adaptations based on a causal graph,
and optimal transport.
}
\examples{
# Simulate data
sim_dat <- simul_dataset()
# Causal graph:
variables <- c("S", "X1", "X2", "Y")
adj <- matrix(
  # S  X1 X2 Y
  c(0, 1, 1, 1,# S
    0, 0, 1, 1,# X1
    0, 0, 0, 1,# X2
    0, 0, 0, 0  # Y
  ),
  ncol = length(variables),
  dimnames = rep(list(variables), 2),
  byrow = TRUE
)
# To visualize the causal graph:
# causal_graph <- fairadapt::graphModel(adj)
# plot(causal_graph)

# Sequential transport according to the causal graph
transported <- seq_trans(data = sim_dat, adj = adj, s = "S", S_0 = 0, y = "Y")
transported
# Transported values from S=0 to S=1, using the causal graph.
transported_val <- as.data.frame(transported$transported)
# Transport of two new observations
new_obs <- data.frame(X1 = c(-2, -1), X2 = c(-1, -2), S = c(0, 0))
transport_new <- seq_trans_new(
  x = transported, newdata = new_obs, data = sim_dat
)
new_obs_transported <- as.data.frame(transport_new)
}
\author{
Fernandes Machado, Agathe and Charpentier, Arthur and Gallic, Ewen.
}
