% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_dataset}
\alias{split_dataset}
\title{Define the function to split the dataset into train/test sets}
\usage{
split_dataset(data, seed, train_ratio = 0.7)
}
\arguments{
\item{data}{dataset}

\item{seed}{seed number to use for replication}

\item{train_ratio}{ratio of data to use in the train set (default to 0.7)}
}
\value{
A list with two elements:
\itemize{
\item \code{data_train}: The train set.
\item \code{data_test}: The test set.
}
}
\description{
Define the function to split the dataset into train/test sets
}
